Use 8 bit control field

+++ src/control.h
@@ -2,5 +2,5 @@
 #include<inttypes.h>
 #include<stdbool.h>
 
-uint16_t control_field(uint8_t send_seq, uint8_t recieve_seq, bool poll);
+uint8_t control_field(uint8_t send_seq, uint8_t recieve_seq, bool poll);
